requirements.txt
anyio==3.6.2
certifi @ file:///croot/certifi_1665076670883/work/certifi
click==8.1.3
dnspython==2.2.1
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.2.0
keras==2.11.0
numpy==1.23.4
paho-mqtt==1.6.1
pandas==1.5.1
pydantic==1.10.2
pymongo==4.3.2
pyotp==2.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
scikit-learn==1.1.2
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.21.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
